Add new interface for networking.
The new interface is plugin based and split into a networkmanager plugin
and socket plugins.
The networkmanager is responsible for processing any networking events
and calling the socket activity functions if activity is detected.
The sockets perform sending and receiving operations.
A socket has callbacks to handle the different kinds of possible events.